% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartage.R
\name{female}
\alias{female}
\title{to calculate the heartage}
\usage{
female(age, hdl, tc, sbp, treat, smoker, DM)
}
\arguments{
\item{age}{numeric}

\item{hdl}{numeric}

\item{tc}{numeric}

\item{sbp}{numeric}

\item{treat}{chr}

\item{smoker}{chr}

\item{DM}{chr}
}
\value{
heartage(gender, point)
}
\description{
to calculate the heartage
}
