% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartage.R
\name{heartage_gender}
\alias{heartage_gender}
\title{to calculate the heartage}
\usage{
heartage_gender(gender, age, hdl, tc, sbp_not, sbp_treat, smoker, DM)
}
\arguments{
\item{gender}{chr}

\item{age}{numeric}

\item{hdl}{numeric}

\item{tc}{numeric}

\item{smoker}{chr}

\item{DM}{chr}

\item{sbp}{numeric}

\item{treat}{chr}
}
\description{
to calculate the heartage
}
