% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartage.R
\name{point}
\alias{point}
\title{Calculate the CVD risk score using Framingham model}
\usage{
point(Gender, Age, HDL, TC, SBP, Treat, Smoker, diabetes)
}
\arguments{
\item{Gender}{chr}

\item{Age}{num}

\item{HDL}{num}

\item{TC}{num}

\item{SBP}{num}

\item{Treat}{chr}

\item{Smoker}{chr}

\item{diabetes}{chr}
}
\value{
numeric The 10-year CVD risk score in percentage.
}
\description{
This function calculates the 10-year cardiovascular disease (CVD) risk score
using the Framingham model. The model is based on the following parameters:
}
\details{
* Age
 * Gender (male/female)
 * HDL cholesterol (HDL)
 * Total cholesterol (TC)
 * Systolic blood pressure (SBP)
 * Whether the patient is receiving treatment for high blood pressure (treat)
 * Whether the patient is a smoker (smoker)
 * Whether the patient has diabetes (DM)
}
\examples{
point("female",61,47,180,124,NA,"yes","no")
df <- data.frame(
        Gender = c("male", "female", "male","female","male"),
        Age = c(55, 60, 65,61,53),
        HDL = c(50, 60, 70,47,55),
        TC = c(200, 220, 240,180,161),
        SBP = c(120, 130, 140,124,125),
        Treat = c("yes", "no", "yes","no","yes"),
        Smoker = c("yes", "no", "yes","yes","no"),
        DM = c("yes", "no", "yes","no","yes")
      )
df \%>\% dplyr::mutate(score = point(Gender, Age, HDL, TC, SBP, Treat, Smoker, DM))

}
