% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartage.R
\name{point_heartage}
\alias{point_heartage}
\title{to calculate the heartage}
\usage{
point_heartage(gender, age, hdl, tc, sbp, treat, smoker, DM)
}
\arguments{
\item{gender}{chr}

\item{age}{numeric}

\item{hdl}{numeric}

\item{tc}{numeric}

\item{sbp}{numeric}

\item{treat}{chr}

\item{smoker}{chr}

\item{DM}{chr}
}
\description{
to calculate the heartage
}
